java streams

+++ BuilderPattern/src/com/examples/java8/Person.java
@@ -6,7 +6,7 @@ public class Person {
 	
 	int id;
 	String name;
-    double salary;
+    int salary;
     Department department;
 	public int getId() {
 		return id;
@@ -23,7 +23,7 @@ public class Person {
 	public double getSalary() {
 		return salary;
 	}
-	public void setSalary(double salary) {
+	public void setSalary(int salary) {
 		this.salary = salary;
 	}
 	public Department getDepartment() {
@@ -32,7 +32,7 @@ public class Person {
 	public void setDepartment(Department department) {
 		this.department = department;
 	}
-	public Person(int id, String name, double salary, Department department) {
+	public Person(int id, String name, int salary, Department department) {
 		super();
 		this.id = id;
 		this.name = name;
